+++ pyproject.toml
@@ -174,6 +174,7 @@ include = [
   "./apps/prairielearn/elements",
   "./apps/prairielearn/python",
   "./contrib/question_converters/canvas",
+  "./scripts/build_images",
   "./graders",
 ]
 exclude = [
